+++ njpeg.h
@@ -15,8 +15,8 @@
 extern int  open_jpeg(const char *,const char *, unsigned int, unsigned int);
 extern void close_jpeg(void);
 extern unsigned int  getpixel_jpeg(int,int,int);
-extern int getwidth_jpeg(void);
-extern int getheight_jpeg(void);
+extern size_t getwidth_jpeg(void);
+extern size_t getheight_jpeg(void);
 
 #endif
